refactor: rename directory ui to parts
